# public/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: 'Complete API documentation for managing Books and Movies with CRUD operations, file uploads, and advanced filtering capabilities.'
  version: 1.0.0
servers:
  -
    url: 'https://rox-test.test'
...
paths:
...
  /api/reset-password:
    post:
      summary: 'Reset password'
      operationId: resetPassword
      description: 'Reset user password using the token received via email. The token is sent when requesting a password reset.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Password has been reset successfully'
                properties:
                  message:
                    type: string
                    example: 'Password has been reset successfully'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid or expired token'
                properties:
                  message:
                    type: string
                    example: 'Invalid or expired token'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: 'The password reset token from email'
                  example: abc123xyz
                  nullable: false
                email:
                  type: string
                  description: "The user's email address"
                  example: john@example.com
                  nullable: false
                password:
                  type: string
                  description: 'The new password (minimum 8 characters)'
                  example: newpassword123
                  nullable: false
                password_confirmation:
                  type: string
                  description: 'Password confirmation (must match password)'
                  example: newpassword123
                  nullable: false
              required:
                - token
                - email
                - password
                - password_confirmation
      security: []
...